Add files via upload
python networking. server script to connect to client machines upload stat script in temp folder and run the command. The client script encrypt the results . Server script decrypt the result and insert the values in a postgresql database which can be used in a rest api or directly called upon in another program. The stat is sent through mail along with alert notification if the statistics (memory usage or cpu usage) crosses a threshold.